pyproject.toml
# file: pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phoneint"
version = "0.1.0"
description = "Phone Number OSINT (lawful/ethical use only)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "phoneint contributors"}]
dependencies = [
  "phonenumbers>=9.0.0",
  "click>=8.1.0",
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gui = [
  "PySide6>=6.6.0",
  "qasync>=0.27.0",
]
pdf = [
  "reportlab>=4.0.0",
]
dev = [
  "black>=24.0.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "types-PyYAML>=6.0.12.20240311",
]

[project.scripts]
phoneint = "phoneint.cli:main"

[tool.setuptools.packages.find]
include = ["phoneint*"]

[tool.setuptools.package-data]
phoneint = ["py.typed", "data/*.json"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
strict_equality = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
  "phonenumbers.*",
  "yaml.*",
  "dotenv.*",
  "PySide6.*",
  "qasync.*",
  "reportlab.*",
]
ignore_missing_imports = true
